+++ tests/test_tool.py
@@ -13,30 +13,3 @@ def test_unpack(wheel_paths, tmpdir):
     """
     for wheel_path in wheel_paths:
         tool.unpack(wheel_path, str(tmpdir))
-
-
-def test_keygen():
-    def get_keyring():
-        WheelKeys, keyring = tool.get_keyring()
-
-        class WheelKeysTest(WheelKeys):
-            def save(self):
-                pass
-
-        class keyringTest:
-            @classmethod
-            def get_keyring(cls):
-                class keyringTest2:
-                    pw = None
-
-                    def set_password(self, a, b, c):
-                        self.pw = c
-
-                    def get_password(self, a, b):
-                        return self.pw
-
-                return keyringTest2()
-
-        return WheelKeysTest, keyringTest
-
-    tool.keygen(get_keyring=get_keyring)